+++ JoysOfEfficiency/JOEMenu.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using JoysOfEfficiency.OptionsElements;
 using JoysOfEfficiency.Utils;
 using Microsoft.Xna.Framework;
@@ -163,6 +164,10 @@ namespace JoysOfEfficiency
                 tab.AddOptionsElement(new LabelComponent("Auto Shearing and Milking"));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoShearingAndMilking", 35, ModEntry.Conf.AutoShearingAndMilking, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedSlider("AnimalHarvestRadius", 14, ModEntry.Conf.AnimalHarvestRadius, 1, 3, OnSliderValueChanged, () => !ModEntry.Conf.AutoShearingAndMilking || ModEntry.Conf.BalancedMode));
+                
+                tab.AddOptionsElement(new EmptyLabel());
+                tab.AddOptionsElement(new LabelComponent("Collect Letter Attachments And Quests"));
+                tab.AddOptionsElement(new ModifiedCheckBox("CollectLetterAttachmentsAndQuests", 36, ModEntry.Conf.CollectLetterAttachmentsAndQuests, OnCheckboxValueChanged));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 _tabs.Add(tab);
@@ -214,9 +219,7 @@ namespace JoysOfEfficiency
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Pause When Idle"));
                 tab.AddOptionsElement(new ModifiedCheckBox("PauseWhenIdle", 33, ModEntry.Conf.PauseWhenIdle, OnCheckboxValueChanged));
-                tab.AddOptionsElement(new ModifiedSlider("IdleTimeout", 12, ModEntry.Conf.IdleTimeout, 1, 300, OnSliderValueChanged, ()=>!ModEntry.Conf.PauseWhenIdle, (which, value) => value + "s"));
-
-
+                tab.AddOptionsElement(new ModifiedSlider("IdleTimeout", 12, ModEntry.Conf.IdleTimeout, 1, 300, OnSliderValueChanged, () => !ModEntry.Conf.PauseWhenIdle, (which, value) => value + "s"));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 _tabs.Add(tab);
@@ -227,11 +230,11 @@ namespace JoysOfEfficiency
 
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Config Menu"));
-                tab.AddOptionsElement(new ModifiedInputListener(this, "KeyShowMenu", 0, ModEntry.Conf.KeyShowMenu, translation, OnInputListnerChanged, OnStartListening));
+                tab.AddOptionsElement(new ModifiedInputListener(this, "keyboard", 0, ModEntry.Conf.ButtonShowMenu, translation, OnInputListnerChanged, OnStartListening));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Auto Harvest"));
-                tab.AddOptionsElement(new ModifiedInputListener(this, "KeyToggleBlackList", 1, ModEntry.Conf.KeyToggleBlackList, translation, OnInputListnerChanged, OnStartListening));
+                tab.AddOptionsElement(new ModifiedInputListener(this, "keyboard", 1, ModEntry.Conf.ButtonToggleBlackList, translation, OnInputListnerChanged, OnStartListening));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 _tabs.Add(tab);
@@ -242,15 +245,15 @@ namespace JoysOfEfficiency
             _isListening = true;
             _listener = option;
         }
-        private void OnInputListnerChanged(int index, Keys value)
+        private void OnInputListnerChanged(int index, SButton value)
         {
             if (index == 0)
             {
-                ModEntry.Conf.KeyShowMenu = value;
+                ModEntry.Conf.ButtonShowMenu = value;
             }
             else if (index == 1)
             {
-                ModEntry.Conf.KeyToggleBlackList = value;
+                ModEntry.Conf.ButtonToggleBlackList = value;
             }
             _mod.WriteConfig();
             _isListening = false;
@@ -293,6 +296,7 @@ namespace JoysOfEfficiency
                 case 33: ModEntry.Conf.PauseWhenIdle = value; break;
                 case 34: ModEntry.Conf.AutoPickUpTrash = value; break;
                 case 35: ModEntry.Conf.AutoShearingAndMilking = value; break;
+                case 36: ModEntry.Conf.CollectLetterAttachmentsAndQuests = value; break;
                 default: return;
             }
             _mod.WriteConfig();
@@ -380,6 +384,14 @@ namespace JoysOfEfficiency
 
         public override void receiveGamePadButton(Buttons b)
         {
+            if (_isListening)
+            {
+                foreach (ModifiedInputListener element in _tabs[_tabIndex].GetElements().OfType<ModifiedInputListener>())
+                {
+                    element.receiveButtonPress(b);
+                }
+                return;
+            }
             if (b.HasFlag(Buttons.DPadUp) && _upCursor.visible)
             {
                 UpCursor();
@@ -496,7 +508,7 @@ namespace JoysOfEfficiency
                 int x2 = (Game1.viewport.Width - 400) / 2;
                 drawTextureBox(b, x2, yPositionOnScreen - 108, 400, 100, Color.White);
 
-                const string str = "JOE Settings";
+                const string str = "JoE Settings";
                 Vector2 size = Game1.dialogueFont.MeasureString(str) * 1.1f;
 
                 Utility.drawTextWithShadow(b, str, Game1.dialogueFont, new Vector2((Game1.viewport.Width - size.X) / 2, yPositionOnScreen - 50 - (int)size.Y / 2), Color.Black, 1.1f);
@@ -580,7 +592,7 @@ namespace JoysOfEfficiency
             {
                 CloseMenu();
             }
-            else if (key == ModEntry.Conf.KeyShowMenu)
+            else if (key.ToSButton() == ModEntry.Conf.ButtonShowMenu)
             {
                 if (!_isFirstTime)
                 {